+++ tests/test_batch.py
@@ -38,7 +38,7 @@ from pasqal_cloud.errors import (
 from pasqal_cloud.utils.constants import BatchStatus, JobStatus
 from pasqal_cloud.utils.filters import BatchFilters
 
-from tests.conftest import mock_core_response
+from tests.conftest import mock_service_response
 from tests.test_doubles.authentication import FakeAuth0AuthenticationSuccess
 from tests.utils import build_query_params, mock_500_http_error_response
 
@@ -46,7 +46,7 @@ from tests.utils import build_query_params, mock_500_http_error_response
 class TestBatch:
     @pytest.fixture
     def load_mock_batch_json_response(self) -> Dict[str, Any]:
-        with open("tests/fixtures/api/v1/batches/_.POST.json") as f:
+        with open("tests/fixtures/core-fast/api/v1/batches/_.POST.json") as f:
             return json.load(f)
 
     @pytest.fixture(autouse=True)
@@ -213,7 +213,6 @@ class TestBatch:
         assert batch.ordered_jobs[0].id == self.job_id
         assert batch.ordered_jobs[0].result == self.job_result
         assert batch.ordered_jobs[0].full_result == self.job_full_result
-        # Ticket (#704)
         with pytest.warns(
             DeprecationWarning,
             match="'jobs' attribute is deprecated, use 'ordered_jobs' instead",
@@ -291,7 +290,7 @@ class TestBatch:
 
         def custom_get_batch_mock(request, _):
             nonlocal call_count
-            resp = mock_core_response(request)
+            resp = mock_service_response(request, "core-fast")
             # Override with custom status the fixture data
             resp["data"]["jobs"][0]["status"] = job_statuses[call_count][0]
             resp["data"]["jobs"][1]["status"] = job_statuses[call_count][1]
@@ -356,7 +355,6 @@ class TestBatch:
     def test_batch_close_and_wait_for_results(
         self, batch: Batch, mock_request: requests_mock.mocker.Mocker
     ):
-        """TODO"""
         batch.close(wait=True)
         assert batch.complete
         assert not batch.open
@@ -364,7 +362,7 @@ class TestBatch:
         assert (
             mock_request.last_request.url
             == f"{self.sdk._client.endpoints.core}/api/v2/jobs?batch_id={batch.id}"
-            "&order_by=ordered_id&order_by_direction=ASC"
+            "&order_by=creation_order&order_by_direction=ASC"
         )
         assert batch.ordered_jobs[0].batch_id == batch.id
         assert batch.ordered_jobs[0].result == self.job_result
@@ -634,7 +632,7 @@ class TestBatch:
         mock_request.reset_mock()
 
         def inline_mock_batch_response(request, _):
-            resp = mock_core_response(request)
+            resp = mock_service_response(request, "core-fast")
             resp["data"]["complete"] = False
             resp["data"]["open"] = True
             resp["data"]["id"] = self.batch_id